# api_schema/openapi-schema.json
{
    "swagger": "2.0",
    "info": {
        "title": "SMERP API",
        "description": "Auto-generated docs for SMERP API",
        "version": "v1"
    },
    "basePath": "/api/v1",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "Basic": {
            "type": "basic"
        }
    },
    "security": [
        {
            "Basic": []
        }
    ],
    "paths": {
...
        "/variable-costs/{id}/delete/": {
            "post": {
                "operationId": "variable-costs_delete",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/VariableCost"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/VariableCost"
                        }
                    }
                },
                "tags": [
                    "variable-costs"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "A unique integer value identifying this Değişken Maliyet.",
                    "required": true,
                    "type": "integer"
                }
            ]
        },
        "/variable-costs/{id}/recover/": {
            "post": {
                "operationId": "variable-costs_recover",
                "description": "",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/VariableCost"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/VariableCost"
                        }
                    }
                },
                "tags": [
                    "variable-costs"
                ]
            },
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "description": "A unique integer value identifying this Değişken Maliyet.",
                    "required": true,
                    "type": "integer"
                }
            ]
        }
    },
    "definitions": {
...
        "VariableCost": {
            "required": [
                "material",
                "material_internal_code",
                "cost"
            ],
            "type": "object",
            "properties": {
                "id": {
                    "title": "ID",
                    "type": "integer",
                    "readOnly": true
                },
                "material": {
                    "title": "Malzeme",
                    "type": "integer"
                },
                "material_internal_code": {
                    "title": "Material internal code",
                    "type": "string",
                    "minLength": 1
                },
                "cost": {
                    "title": "Cost",
                    "type": "number",
                    "format": "decimal"
                },
                "uom": {
                    "title": "Birim",
                    "type": "string",
                    "enum": [
                        "ADT",
                        "KG",
                        "G",
                        "L",
                        "ML",
                        "M",
                        "BOX",
                        "PLT"
                    ]
                },
                "source_type": {
                    "title": "Source type",
                    "type": "string",
                    "readOnly": true
                },
                "procurement_order": {
                    "title": "Satınalma Siparişi",
                    "type": "integer",
                    "readOnly": true,
                    "x-nullable": true
                },
                "user": {
                    "title": "Kullanıcı",
                    "type": "integer",
                    "x-nullable": true
                },
                "created_by": {
                    "title": "Created by",
                    "type": "string",
                    "readOnly": true
                },
                "created_at": {
                    "title": "Created at",
                    "type": "string",
                    "readOnly": true
                },
                "source_object": {
                    "title": "Source object",
                    "type": "string",
                    "readOnly": true
                }
            }
        }
    }
}